Rename /controllers to /domain
